New extra step for wavenumbers
The path finder for the wavenumbers does quickly finds lots of them, but can often leave some fall through the cracks. This is now finally corrected by adding an extra step which meshes over all the wavenumbers already found to make sure nothing is missing. Using constrained optim was essential for this.