Change how effective wavenumbers travelling in the wrong direction are filtered.
the criteria now checks if  -low_tol < abs(imag(K))/real(K) < 0, i.e. that is travelling in the wrong direction with very small attenuation.
This came about from finding a K with imag(K) < 0 and real(K) >0, that was correct (gives the right reflection coefficient and matches other methods), but slightly unphysical